ASoC: Intel: debugfs: remove sst_debugfs_init/exit - part 1
These functions merely create and remove debugfs directories. Given
this simplicity, there is really no need to have dedicated functions.
While there, improve logic to create the debugfs root only when
missing.

In order to not break build or functionality. The sst_debugfs_init
funciton is kept but only to see the debugfs rootdir. This will be
removed in a subsequent patch.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>