ASoC: Intel: debugfs: simplify file creation/deletion
Simplify the signature of the functions that create/delete files to
expose the mmio buffers. This will present a simpler interface to the
users of debugfs.

On the downside, sst-debugfs may become too specific to the DSP. Also,
names of the files become more restricted.

sst-dsp.h was added as sst-dsp-priv.h require a forward declaration
present in the former.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>