ASoc: Intel: HSW/BDW: Implement FW log capabilities
The firmware running in the Intel audio DSP is capable of generating
logs about the activity in the DSP. Such logs are transmitted to the
host processor via DMA transfers.

In order to enable log info from the firmware, the host processor
must issue a IPC_DEBUG_ENABLE_LOG message. Then, the DSP will start
filling a DMA buffer. When such buffer is full, the DSP will issue
a IPC_DEBUG_REQUEST_LOG_DUMP message for the host processor to flush
the buffer.

Once the buffer has been flushed, the host processor will
notify the DSP about the new position on which it should continue
working in the DMA buffer by issuing a IPC_DEBUG_NOTIFY_LOG_DUMP
message along with the position.

Finally, when the host processor no longer desires log info from the
firmware, it will simply issue a IPC_DEBUG_DISABLE_LOG message to
the DSP.

The log info capture infrastructure is based debugfs. debugfs will
create a fw/log file to which the DMA buffer will be flushed. The
IPC_DEBUG_ENABLE_LOG and IPC_DEBUG_DISABLE_LOG messages will be sent
to the DSP when opening and closing the debugfs file, respectively.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>