ASoC: Intel: HSW/BDW: add alsa DMA buffer descriptors
The firmware in the Intel audio DSP is capable of generating a firmware
log. Such log is passed to the host processor via DMA transfers. Alsa
provides descriptors for DMA buffers.

In order to keep track of such buffers during the lifetime of the
driver, add two descriptors to the log stream structure.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>